requirements.txt
fastapi
uvicorn
requests
Pillow
torch
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
